feat: Add `--latest` option flag to `dvm releases` (#57)
* feat: use constant for 'channel' option key in ReleasesCommand

This commit introduces a constant `_channelKey` to represent the 'channel' option key within the `ReleasesCommand` class. This change ensures consistency and reduces the likelihood of errors due to typos in string literals. The constant is used both when adding the option to the argument parser and when retrieving the argument result, enhancing maintainability and readability of the code.

* feat: add 'latest' flag to ReleasesCommand for filtering releases

This commit introduces a new flag, represented by the constant `_latestKey`, to the `ReleasesCommand` class. This flag allows users to filter the output to show only the latest release. The flag is not negatable, emphasizing its use as a boolean switch. This enhancement improves the command's flexibility, enabling users to quickly access the most recent release information without parsing through the entire list of releases.

* feat: implement 'latest' release filtering in ReleasesCommandService

This commit enhances the `ReleasesCommandService` by adding support for filtering to show only the latest SDK release. It introduces a new parameter `isLatest` to the service call method, which, when set to `true`, results in the service outputting only the most recent release information. This functionality is integrated with the `ReleasesCommand` to provide users with an option to view just the latest release details, improving the command's utility for users interested in the most current release without the need to sift through all available versions.