feat: support `latest` option in SDK install (#58)
* feat: add method to fetch latest SDK version in SdkService

This commit introduces a new method `getLatestSdk` in the `SdkService` class, designed to fetch the latest SDK version for a specified channel. The method retrieves all SDK versions for the given channel and returns the most recent one. If no versions are found, it throws an exception, ensuring that callers are aware of the failure to locate any Dart SDK versions for the provided channel. This addition enhances the service's capabilities by allowing more precise retrieval of SDK information, catering to use cases that require only the latest version details.

* refactor: optimize latest SDK retrieval in ReleasesCommandService

This commit refines the `ReleasesCommandService` logic for fetching the latest SDK version. Instead of retrieving all SDK versions and then selecting the last one, the service now directly calls the newly added `getLatestSdk` method from the `SdkService` when the `isLatest` flag is true. This approach not only simplifies the code but also improves efficiency by eliminating unnecessary retrieval and processing of all SDK versions when only the latest is required. This change ensures a more direct and resource-efficient way to obtain the latest SDK version, aligning with best practices for performance and code clarity.

* feat: extend `InstallCommand` with options to install the latest SDK release and specify the channel in `install_command.dart`

* feat: support `latest` option in SDK install

This commit introduces functionality to handle the `latest` option when installing SDK versions. A new `LatestOptions` type and an `sdkChannel` parameter have been added to support this feature, allowing the installation of the latest SDK version from a specified channel. This update enhances the `InstallCommandService` and `InstallCommand` classes to accommodate the latest SDK installation process, improving the tool's flexibility and user experience in managing SDK versions.

* feat(install_command): expand help text for latest release flag

* refactor: standardize help text position in command args

* style: standardize punctuation in help texts and invocations